feat: adjust sidebar to include chats
