+++ package.json
@@ -22,6 +22,7 @@
     "lefthook": "^1.13.1",
     "lint-staged": "^16.1.6",
     "lucide-react": "^0.544.0",
+    "markdown-to-jsx": "^7.7.13",
     "next": "15.5.3",
     "pdf-parse": "^1.1.1",
     "prettier": "^3.6.2",
