+++ src/features/shared/components/sidebar/link-item.tsx
@@ -14,7 +14,7 @@ export function LinkItem({
     <Link
       href={href}
       aria-current={active ? "page" : undefined}
-      className={`relative flex items-center gap-3 py-2 px-0 justify-center transition-all duration-200 text-black hover:bg-muted group-hover:px-3 group-hover:justify-start ${active ? "bg-muted" : ""}`}
+      className={`relative flex items-center gap-3 py-2 px-0 justify-center transition-all duration-200 text-black hover:bg-muted group-hover:px-3 group-hover:justify-start ${active ? "border-primary/40 bg-primary/10" : "border-transparent hover:bg-muted"}`}
     >
       <span
         className={`absolute left-0 top-0 h-full w-0.5 transition-colors ${active ? "bg-primary" : "bg-transparent"}`}
