+++ src/features/shared/components/sidebar/sidebar-shell-shared.tsx
@@ -1,0 +1,5 @@
+"use client";
+
+export function SidebarDivider() {
+  return <div className="mx-2 my-2 h-px bg-border" />;
+}
